refactor(api): extract pairing handlers into routes/pairing.rs
routes/system.rs grew to 5897 lines spanning 8 unrelated domains. Move
the device pairing sub-domain (request, complete, devices, notify,
remove_device + the resolve_pairing_base_url helper and pairing_tests)
into a sibling module routes/pairing.rs and merge it back into
system::router(). Public route paths under /pairing/* are unchanged;
glob re-export via routes::* keeps openapi.rs path bindings working.

Refs #3749 (1 of 8 sub-domain extractions).
