+++ crates/librefang-api/src/routes/mod.rs
@@ -28,6 +28,7 @@ pub mod mcp_auth;
 pub mod media;
 pub mod memory;
 pub mod network;
+pub mod pairing;
 pub mod plugins;
 pub mod prompts;
 pub mod providers;
@@ -64,6 +65,7 @@ pub use mcp_auth::*;
 pub use media::*;
 pub use memory::*;
 pub use network::*;
+pub use pairing::*;
 pub use plugins::*;
 pub use providers::*;
 pub use skills::*;
